Split //test/extensions/filters/http/ext_authz:ext_authz_test
