+++ test/extensions/filters/http/ext_authz/BUILD
@@ -15,12 +15,78 @@ licenses(["notice"])  # Apache 2
 
 envoy_package()
 
+envoy_extension_cc_test_library(
+    name = "http_filter_test_base_lib",
+    hdrs = ["http_filter_test_base.h"],
+    extension_names = ["envoy.filters.http.ext_authz"],
+    deps = [
+        "//envoy/common:optref_lib",
+        "//envoy/http:filter_interface",
+        "//envoy/http:metadata_interface",
+        "//envoy/http:query_params_interface",
+        "//envoy/network:address_interface",
+        "//envoy/network:connection_interface",
+        "//envoy/stats:scope_interface",
+        "//envoy/stream_info:stream_info_interface",
+        "//envoy/tracing:trace_driver_interface",
+        "//source/common/buffer:buffer_lib",
+        "//source/common/http:headers_lib",
+        "//source/common/network:address_lib",
+        "//source/common/protobuf",
+        "//source/extensions/filters/common/ext_authz:ext_authz_interface",
+        "//source/extensions/filters/http/ext_authz",
+        "//test/extensions/filters/common/ext_authz:ext_authz_mocks",
+        "//test/mocks/http:http_mocks",
+        "//test/mocks/network:connection_mocks",
+        "//test/mocks/server:server_factory_context_mocks",
+        "//test/mocks/stats:stats_mocks",
+        "//test/mocks/upstream:cluster_manager_mocks",
+        "//test/test_common:utility_lib",
+        "@envoy_api//envoy/extensions/filters/http/ext_authz/v3:pkg_cc_proto",
+        "@envoy_api//envoy/service/auth/v3:pkg_cc_proto",
+    ],
+)
+
 envoy_extension_cc_test(
     name = "ext_authz_test",
     srcs = ["ext_authz_test.cc"],
     extension_names = ["envoy.filters.http.ext_authz"],
     rbe_pool = "6gig",
     deps = [
+        ":http_filter_test_base_lib",
+        "//envoy/http:codes_interface",
+        "//source/common/buffer:buffer_lib",
+        "//source/common/common:empty_string",
+        "//source/common/http:context_lib",
+        "//source/common/http:headers_lib",
+        "//source/common/json:json_loader_lib",
+        "//source/common/network:address_lib",
+        "//source/common/protobuf:utility_lib",
+        "//source/extensions/filters/common/ext_authz:ext_authz_grpc_lib",
+        "//source/extensions/filters/http/ext_authz",
+        "//test/extensions/filters/common/ext_authz:ext_authz_mocks",
+        "//test/mocks/buffer:buffer_mocks",
+        "//test/mocks/http:http_mocks",
+        "//test/mocks/network:network_mocks",
+        "//test/mocks/server:server_factory_context_mocks",
+        "//test/mocks/upstream:cluster_manager_mocks",
+        "//test/proto:helloworld_proto_cc_proto",
+        "//test/test_common:test_runtime_lib",
+        "//test/test_common:utility_lib",
+        "@envoy_api//envoy/config/core/v3:pkg_cc_proto",
+        "@envoy_api//envoy/extensions/filters/http/ext_authz/v3:pkg_cc_proto",
+        "@envoy_api//envoy/service/auth/v3:pkg_cc_proto",
+        "@envoy_api//envoy/type/v3:pkg_cc_proto",
+    ],
+)
+
+envoy_extension_cc_test(
+    name = "ext_authz_filter_test",
+    srcs = ["ext_authz_filter_test.cc"],
+    extension_names = ["envoy.filters.http.ext_authz"],
+    rbe_pool = "6gig",
+    deps = [
+        ":http_filter_test_base_lib",
         "//envoy/http:codes_interface",
         "//source/common/buffer:buffer_lib",
         "//source/common/common:empty_string",
